sound: Fix frequency handling and add RPM rev test (#29)
- Change curr_frequency from uint32_t to double for precise frequency
  control
- Remove get_closest_frequency() and change phase calculation in engine
  sound to ensure continuous sine curve
- Add RPM revving test

+++ src/monocoque/devices/sounddevice.h
@@ -26,7 +26,7 @@ typedef struct
     uint32_t amplitude;
     uint32_t amplitudeMax;
     double duration;
-    uint32_t curr_frequency;
+    double curr_frequency;
     uint32_t curr_amplitude;
     double curr_duration;
     double phase;
